realTimeTranslate/requirements.txt
Flask
google-cloud-speech
google-cloud-translate
pyaudio
